==> working/1_prepare_data.ipynb <==
# Setup

%load_ext autotime

import pickle
import zipfile

import numpy as np
import pandas as pd

from tqdm import tqdm_notebook

# Targets & IDs

Load the train names (`Id`) and labels (`Attack`):

df_train = pd.read_csv('../input/ytrain.csv')
print("Train shape:", df_train.shape)
df_train.head()

Load the sample submission for test names (`Id`):

df_test = pd.read_csv('../input/SampleSubmission.csv').drop(columns="Attack")
print("Test shape:", df_test.shape)
df_test.head()

Concatenate train and test:

df = pd.concat(
    [df_train, df_test],
    ignore_index=True,
)[["Id", "Attack"]]

print("Total shape:", df.shape)

df.head()

df.tail()

names = df.Id.values

y = df.Attack.values

# Time series

Load time series and shape them to be a 3D-tensor - with dimensions `(observations, series, points)`:

series = {}

for archive in '../input/xtrain.zip', '../input/xtest.zip':
    with zipfile.ZipFile(archive) as zf:
        for name in tqdm_notebook(zf.namelist(), desc=archive.split("/")[-1]):
            if name.endswith('.csv'):
                
                fname = name.split('/')[-1]
                Id = int(fname.split('.')[0])

                df = pd.read_csv(zf.open(name), header=None)
                series[Id] = df.values

X = np.array([series[name] for name in names]).transpose((0, 2, 1))
print("X shape:", X.shape)

So there's a total of 500 observations, 56 time series for each observation and a total of 5760 time points for each time series.

# Save

Pickle the data for later use:

data = {
    "X": X,
    "y": y,
    "names": names,
}

with open("data.pickle", "wb") as f:
    pickle.dump(data, f)

!ls -lh data.pickle
