==> working/3_synthetic_dataset.ipynb <==
# Setup

%load_ext autotime

import pickle

import numpy as np
from tqdm import tqdm_notebook

# Load the data

with open("data.pickle", "rb") as f:
    data = pickle.load(f)

X = data["X"][:, 2:, :]
y = data["y"]
names = data["names"]

N.b.: first two series of each observation are unused, as I found no useful patterns in them.

# Generate synthetic dataset

## Negative observations

(i.e. without anomalies)

Selecting only clean (non-attacked) factories:

i_neg = np.where(y == 0)[0]
X_neg = np.vstack(X[i_neg])

Cut into chunks:

def generate_chunks(l, o):
    chunks = []
    start = 0
    while start + l < X.shape[-1]:
        chunks.append((start, start + l))
        start += (l - o)
    return chunks

CHUNK_LENGTH = 1000
CHUNK_OVERLAP = 900

chunks = generate_chunks(CHUNK_LENGTH, CHUNK_OVERLAP)

X_chunks = np.vstack([
    X_neg[:, chunk[0]:chunk[1]]
    for chunk in chunks
])

neg_size = X_chunks.shape[0]
neg_size

Save the negative part:

file_name_neg = f"chunks_neg_{CHUNK_LENGTH}_{CHUNK_OVERLAP}.pickle"

with open(file_name_neg, "wb") as f:
    pickle.dump(X_chunks, f)

! ls -lh $file_name_neg

## Positive observations

(i.e. with anomalies)

Adding random patterns like ones found by EDA.

### Peaks

def add_peak(x):

    if np.abs(x.mean()) <= x.std():
        v = np.random.uniform(-50, 50, 1)[0]
    else:
        v = 0

    length = np.random.choice([3, 4, 5])
    pos = np.random.choice(range(len(x) - length - 1))

    x[pos:(pos + length)] = v

    return True

### Noises

def add_noise(x):

    size = np.random.uniform(low=10, high=50)
    mean = x.mean()

    if x.std() > 0.01:
        size = size * x.std()

    pos = np.random.choice(range(len(x)))
    width = int(np.random.uniform(low=100, high=800))
    is_start = bool(np.random.binomial(1, 0.5))

    if is_start:
        pos_start = pos
        pos_end = pos + width
    else:
        pos_start = pos - width
        pos_end = pos

    if pos_start < 0:
        pos_start = 0
    if pos_end > len(x):
        pos_end = len(x)

    if pos_end - pos_start < 100:
        return False

    x[pos_start:pos_end] = np.random.uniform(
        mean - size / 2,
        mean + size / 2,
        pos_end - pos_start,
    )

    return True

### Constants

def add_constant(x):

    if np.unique(x).shape[0] == 1:
        return False

    pos = np.random.choice(range(len(x)))
    width = int(np.random.uniform(low=100, high=300))
    is_start = bool(np.random.binomial(1, 0.5))

    if is_start:
        pos_start = pos
        pos_end = pos + width
    else:
        pos_start = pos - width
        pos_end = pos

    if pos_start < 0:
        pos_start = 0
    if pos_end > len(x):
        pos_end = len(x)

    if pos_end - pos_start < 100:
        return False

    x[pos_start:pos_end] = x[pos_start]

    return True

### "Special cases"

def add_special_case(x):

    if np.abs(x.mean()) <= x.std() or x.mean() < 0:
        return False

    pos = np.random.choice(range(len(x)))
    width = int(np.random.uniform(low=300, high=800))
    is_start = bool(np.random.binomial(1, 0.5))

    if is_start:
        pos_start = pos
        pos_end = pos + width
    else:
        pos_start = pos - width
        pos_end = pos

    if pos_start < 0:
        pos_start = 0
    if pos_end > len(x):
        pos_end = len(x)

    if pos_end - pos_start < 100:
        return False

    insert = np.zeros(pos_end - pos_start)
    n_peaks = int(np.random.uniform(10, 20))
    peaks_pos = np.random.choice(range(len(insert)), size=n_peaks)
    insert[peaks_pos] = np.random.uniform(low=0, high=x.mean(), size=len(peaks_pos))

    x[pos_start:pos_end] = insert

    return True

### Randomly apply to chunks

fs = [add_peak, add_noise, add_constant, add_special_case]

for i in tqdm_notebook(range(X_chunks.shape[0])):
    while True:
        f = np.random.choice(fs, size=1)[0]
        if f(X_chunks[i]):
            break

Save the positive part:

file_name_pos = f"chunks_pos_{CHUNK_LENGTH}_{CHUNK_OVERLAP}.pickle"

with open(file_name_pos, "wb") as f:
    pickle.dump(X_chunks, f)

! ls -lh $file_name_pos
