==> working/4_synthetic_model.ipynb <==
# Setup

%load_ext autotime
%matplotlib inline

Hack to load TensorFlow from system libraries:

import sys

sys.path.append("/usr/lib/python3.6/site-packages/")

(I run everything from virtual environment, but there's already built package of TensorFlow with CUDA support in ArchLinux repo, so I use it instead of building myself.)

import tensorflow as tf

Ensure that it will run on GPU:

from tensorflow.python.client import device_lib


def get_available_gpus():
    local_device_protos = device_lib.list_local_devices()
    return [x.name for x in local_device_protos if x.device_type == 'GPU']


available_gpus = get_available_gpus()
print("Available GPUs:", available_gpus)
assert len(available_gpus) > 0

Other usual stuff:

import pickle

import numpy as np
import pandas as pd
from matplotlib import pyplot as plt

import tqdm

# Load the Data

with open("chunks_neg_1000_900.pickle", "rb") as f:
    X_neg = pickle.load(f)

with open("chunks_pos_1000_900.pickle", "rb") as f:
    X_pos = pickle.load(f)

y_neg = np.zeros(X_neg.shape[0])
y_pos = np.ones(X_pos.shape[0])

Full dataset:

X = np.vstack([X_neg, X_pos])
y = np.concatenate((y_neg, y_pos))

Normalize dataset:

def normalize(X):
    X_std = X.std(axis=1, keepdims=True)
    X_mean = X.mean(axis=1, keepdims=True)
    X_std[np.where(X_std == 0)[0]] = 1
    return (X - X_mean) / X_std

X = normalize(X)

Prepare training data:

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1)

Well, "test" is actually "validation".

# The Model

## Build

from keras.layers import (
    Conv1D,
    Dense,
    Flatten,
    Input,
    MaxPooling1D,
    Reshape,
)
from keras.models import Model
from keras.optimizers import Adam

def make_model():
    inputs = Input(shape=X.shape[1:])
    
    x = inputs
    x = Reshape(X.shape[1:] + (-1,))(x)
    x = Conv1D(8, 3, activation="relu")(x)
    x = Conv1D(8, 3, activation="relu")(x)
    x = MaxPooling1D(2)(x)
    x = Conv1D(16, 3, activation="relu")(x)
    x = Conv1D(16, 3, activation="relu")(x)
    x = MaxPooling1D(2)(x)
    x = Conv1D(32, 3, activation="relu")(x)
    x = Conv1D(32, 3, activation="relu")(x)
    x = MaxPooling1D(2)(x)
    x = Flatten()(x)
    x = Dense(100, activation="relu")(x)
    x = Dense(10, activation="relu")(x)
    x = Dense(1, activation="sigmoid")(x)
    outputs = x
    
    model = Model(inputs=inputs, outputs=outputs)
    model.summary()
    
    model.compile(
        loss="binary_crossentropy",
        optimizer=Adam(lr=0.0008),
    )
    return model

## Train

Sample for evaluating performance on a training set:

i_train_sample = np.random.choice(range(X_train.shape[0]), size=X_test.shape[0], replace=False)
X_train_sample = X_train[i_train_sample]
y_train_sample = y_train[i_train_sample]

BATCH_SIZE = 5000
EPOCHS = 22
LOG_EVERY = 1

from sklearn.metrics import roc_auc_score

***N.b.: actually you should not run the next two code cells, they were used to test different architectures and to find out a number of epochs required to reach the optimum.***

Instantiate the model:

```python
m = make_model()
```

Train the model on a part of train set and evaluate both on train and on a validation (called "test" here):

```python
df_log = pd.DataFrame(columns=["epoch", "train_auc", "test_auc"])

for epoch in range(1, EPOCHS + 1, LOG_EVERY):
    m.fit(X_train, y_train, batch_size=BATCH_SIZE, epochs=LOG_EVERY)
    
    yhat_train_sample = m.predict(X_train_sample)
    train_auc = roc_auc_score(y_train_sample, yhat_train_sample)
    
    yhat_test = m.predict(X_test)
    test_auc = roc_auc_score(y_test, yhat_test)
    
    row = (epoch + LOG_EVERY - 1, train_auc, test_auc)
    print("\nEPOCH #{} - train AUC {:.3f} test AUC {:.3f}".format(*row))
    
    df_log.loc[df_log.shape[0]] = row
    
    f, a = plt.subplots(figsize=(10, 6))
    df_log.set_index("epoch").plot(ax=a, marker=".")
    plt.grid()
    plt.show()
```

## Score the real thing

Model to fit on full train:

m = make_model()

Fit on the full train:

m.fit(X, y, batch_size=BATCH_SIZE, epochs=EPOCHS)

Metric on a training set:

roc_auc_score(y, m.predict(X, batch_size=BATCH_SIZE))

Loading the real data:

with open("data.pickle", "rb") as f:
    data_real = pickle.load(f)

X_real = data_real["X"][:, 2:, :]
y_real = data_real["y"]
names_real = data_real["names"]

X_real.shape, y_real.shape, names_real.shape

Treat each of the time series individually:

X_real = np.vstack(X_real)
X_real.shape

Split them into chunks:

def generate_chunks(l, o):
    chunks = []
    start = 0
    while start + l < X_real.shape[-1]:
        chunks.append((start, start + l))
        start += (l - o)
    return chunks


CHUNK_LENGTH = 1000
CHUNK_OVERLAP = 900

chunks = generate_chunks(CHUNK_LENGTH, CHUNK_OVERLAP)

len(chunks)

i_known = np.where((y_real[:, None] == [0, 1]).max(axis=1))[0]
i_unknown = np.where(~(y_real[:, None] == [0, 1]).max(axis=1))[0]

Calculate predicts for each chunk:

yhats_real = []

for chunk in tqdm.tqdm(chunks):
    yhat_real_one = m.predict(normalize(X_real[:, chunk[0]:chunk[1]]))
    yhats_real.append(yhat_real_one)

yhat_real = np.hstack(yhats_real)

Smooth with moving average:

MA_PERIOD = 3

yhat_convolved = np.vstack([
    np.convolve(yhat_real[i], np.ones((MA_PERIOD,)) / MA_PERIOD, mode="valid")
    for i in range(yhat_real.shape[0])
])

yhat_convolved.shape

yhat_reshaped = yhat_convolved.max(axis=1).reshape(y_real.shape[0], -1).max(axis=1)

Evaluate thus created scores on the train dataset:

roc_auc_score(y_real[i_known], yhat_reshaped[i_known])

For submission:

pd.DataFrame({
    "Id": names_real[i_unknown],
    "Attack": yhat_reshaped[i_unknown],
})[["Id", "Attack"]].to_csv("cnn.980.csv", index=False)

For stacking:

df_y = pd.DataFrame({
    "y": y_real,
    "yhat": yhat_reshaped,
})

(
    df_y
    .rename(columns={"y": "Attack", "yhat": "AttackHat"})
    .to_csv("cnn.980.full.csv", index_label="Id")
)
