==> working/5_final_model.ipynb <==
# Setup

%load_ext autotime
%matplotlib inline

import pickle

import numpy as np
import pandas as pd
import scipy as sp

from scipy import stats

# Load the Data

with open("data.pickle", "rb") as f:
    data = pickle.load(f)

X = data["X"][:, 2:, :]
y = data["y"]
names = data["names"]

# Normalize the Data

def normalize(X):
    X_iqr = stats.iqr(X, axis=2, keepdims=True)
    X_iqr = np.where(X_iqr == 0, np.ones_like(X_iqr), X_iqr)
    X_median = np.median(X, axis=2, keepdims=True)
    return ((X - X_median) / X_iqr / 2)

X_normalized = normalize(X)

# NN predictions

df_y = pd.read_csv("cnn.980.full.csv")
df_y.head()

from sklearn.metrics import roc_auc_score

roc_auc_score(
    df_y[~df_y.Attack.isnull()].Attack,
    df_y[~df_y.Attack.isnull()].AttackHat,
)

nn_pred = df_y.AttackHat.as_matrix()

# Statistical features

def make_stat_features(X):
    X_min = X.min(axis=2, keepdims=True)
    X_max = X.max(axis=2, keepdims=True)
    X_mean = X.mean(axis=2, keepdims=True)
    X_ptp = X_max - X_min
    X_std = X.std(axis=2, keepdims=True)
    X_median = np.median(X, axis=2, keepdims=True)
    X_iqr = stats.iqr(X, axis=2, keepdims=True)
    X_mean_diff = np.abs(X - X_mean).max(axis=2, keepdims=True)
    X_median_diff = np.abs(X - X_median).max(axis=2, keepdims=True)

    raw_features = [
        X_min,
        X_max,
        X_mean,
        X_ptp,
        X_std,
        X_median,
        X_iqr,
        X_mean_diff,
        X_median_diff,
    ]

    return np.hstack([
        v.squeeze()
        for v in raw_features
    ])

X_features_1 = make_stat_features(X)

X_features_2 = make_stat_features(X_normalized)

# Model

i_train = np.where((y[:, None] == ([0, 1])).max(axis=1))[0]
i_test = np.where(~(y[:, None] == ([0, 1])).max(axis=1))[0]

from sklearn.model_selection import cross_val_score, cross_val_predict
from sklearn.linear_model import LogisticRegression

### Linear model

Evaluation:

cross_val_score(
    LogisticRegression(),
    X_features_1[i_train],
    y[i_train],
    scoring="roc_auc",
    cv=5,
).mean()

cross_val_score(
    LogisticRegression(),
    X_features_2[i_train],
    y[i_train],
    scoring="roc_auc",
    cv=5,
).mean()

Out-of-fold predictions for train:

lr_pred_1_train = cross_val_predict(
    LogisticRegression(),
    X_features_1[i_train],
    y[i_train],
    cv=5,
    method="predict_proba"
)[:, 1]

lr_pred_2_train = cross_val_predict(
    LogisticRegression(),
    X_features_2[i_train],
    y[i_train],
    cv=5,
    method="predict_proba"
)[:, 1]

Predictions for test:

lr_pred_1_test = LogisticRegression().fit(
    X_features_1[i_train],
    y[i_train],
).predict_proba(X_features_1[i_test])[:, 1]

lr_pred_2_test = LogisticRegression().fit(
    X_features_2[i_train],
    y[i_train],
).predict_proba(X_features_2[i_test])[:, 1]

Concatenate train and test predictions:

lr_pred_1 = np.zeros_like(y)
lr_pred_1[i_train] = lr_pred_1_train
lr_pred_1[i_test] = lr_pred_1_test

lr_pred_2 = np.zeros_like(y)
lr_pred_2[i_train] = lr_pred_2_train
lr_pred_2[i_test] = lr_pred_2_test

### LightGBM as a stacker

from lightgbm import LGBMClassifier

One setup:

lgbm = LGBMClassifier(
    n_estimators=80,
    num_leaves=4,
    reg_alpha=1e-6,
    reg_lambda=0,
)

cross_val_score(
    lgbm,
    np.hstack([
        X_features_1[i_train],
        X_features_2[i_train],
        lr_pred_1[:, None][i_train],
        lr_pred_2[:, None][i_train],
        nn_pred[:, None][i_train],
    ]),
    y[i_train],
    scoring="roc_auc",
    cv=5,
).mean()

lgbm.fit(
    np.hstack([
        X_features_1[i_train],
        X_features_2[i_train],
        lr_pred_1[:, None][i_train],
        lr_pred_2[:, None][i_train],
        nn_pred[:, None][i_train],
    ]),
    y[i_train],
);

lgbm_pred_test = lgbm.predict_proba(
    np.hstack([
        X_features_1[i_test],
        X_features_2[i_test],
        lr_pred_1[:, None][i_test],
        lr_pred_2[:, None][i_test],
        nn_pred[:, None][i_test],
    ]),
)[:, 1]

pd.DataFrame({
    "Id": names[i_test],
    "Attack": lgbm_pred_test,
})[["Id", "Attack"]].to_csv("lgbm.9898.csv", index=False)

Another setup:

lgbm = LGBMClassifier(
    n_estimators=350,
    num_leaves=4,
    reg_alpha=0,
    reg_lambda=0,
)

cross_val_score(
    lgbm,
    np.hstack([
#         X_features_1[i_train],
#         X_features_2[i_train],
        lr_pred_1[:, None][i_train],
        lr_pred_2[:, None][i_train],
        nn_pred[:, None][i_train],
    ]),
    y[i_train],
    scoring="roc_auc",
    cv=5,
).mean()

lgbm.fit(
    np.hstack([
#         X_features_1[i_train],
#         X_features_2[i_train],
        lr_pred_1[:, None][i_train],
        lr_pred_2[:, None][i_train],
        nn_pred[:, None][i_train],
    ]),
    y[i_train],
);

lgbm_pred_test = lgbm.predict_proba(
    np.hstack([
#         X_features_1[i_test],
#         X_features_2[i_test],
        lr_pred_1[:, None][i_test],
        lr_pred_2[:, None][i_test],
        nn_pred[:, None][i_test],
    ]),
)[:, 1]

pd.DataFrame({
    "Id": names[i_test],
    "Attack": lgbm_pred_test,
})[["Id", "Attack"]].to_csv("lgbm.984.csv", index=False)
